% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interplot.medline.R
\name{interplot.medline}
\alias{interplot.medline}
\title{A Wrapper for Interplot (Holt and Su) to add Median Lines}
\usage{
interplot.medline(m, var1, var2, ci = 0.95)
}
\arguments{
\item{m}{A model object with interaction terms}

\item{var1}{The variable for which you wish to calculate marginal effects}

\item{var2}{The moderator}
}
\value{
The function returns a \code{\link{ggplot}} object.
}
\description{
This package adds a median line to a standard
marginal effects plot (Brambor, Clark, and Golder 2006) created by \code{\link{interplot}} (Solt and Hu 2015).
}
\examples{
 interplot.medline(m = fit, var1 = "cyl", var2 = "wt")

}
\references{
Brambor, Thomas, William Roberts Clark and Matt Golder. 2006. Understanding Interaction Models: Improving Empirical Analyses. \emph{Political Analysis} 14(1), pp. 63-82.

Solt, Frederick, and Yue Hu. 2015. \code{interplot}: Plot the Coefficients of Variables in Interaction Terms. The Comprehensive R Archive Network (CRAN).
}
\seealso{
\code{\link{interplot}}
}
