% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interplot.medline.R
\name{interplot.medline}
\alias{interplot.medline}
\title{A Wrapper for Interplot to add Median Lines}
\usage{
interplot.medline(m, var1, var2, ci = 0.95)
}
\arguments{
\item{m}{A model object with interaction terms}

\item{var1}{The variable for which you wish to calculate marginal effects}

\item{var2}{The moderator}
}
\value{
The function returns a \code{ggplot} object.
}
\description{
This package adds a median line to a standard
marginal effects plot
}
\examples{
 interplot.medline(m = mpg, var1 = "cyl", var2 = "wt")

}
